Move record related method into Searchkick::Index::Records
